MXS-4656: Use mxs::Reply to fix transaction state
If the transaction state reported by the server differs from the one in
MaxScale, the state in MaxScale is fixed to be the same.

This commit replaces the old implementation of the transaction tracking
in the MariaDB client protocol code with the new one in the
mariadb::TrxTracker class. The duplicated code should've been removed as
a part of the fix to MXS-4419.

The `session_track_trx_state` is now deprecated as the transaction state
fixing is always done if the module requires transaction tracking.