MXS-4629 Don't mark a column a unique key if it's a part of composite key
The `col-definitions` component shows UQ input for each column as
a checkbox, so it can't handle composite unique key. Tick/untick
that UQ box of a column will add/remove a UQ key for that column.

When dropping a column that has both a UQ key and is part of a
composite unique key, the single UQ key object is deleted
while the composite key is modified instead of deleting. The composite
key will be altered to include the remaining columns in the next
commit.