MXS-4665: Fix listener creation errors
The errors were logged in the MaxScale log instead of being captured and
sent to the REST-API client. Now that the real errors are visible to the
client, the error messages in config_runtime.cc can be removed.

Also fixed the order of the operations that are done when a listener is
about to be started. Previously, the listener configuration was
persisted and then an attempt to start was made. If the listener failed
to start, a possibly faulty configuration was persisted. The correct
thing to do is to only persist the configuration if the listener was
successfully started.