MXS-4618: Offload blocking write to a separate thread
The write into the command's pipe can block and thus needs to be done on
a separate thread. Since the filter itself is not used directly, the new
code doesn't rely on the filter session being alive and the trickery
with std::weak_ptr can be avoided.

The rate of data upload is useful for estimating upload performance and
it should be logged even in the LDLI-to-xpand_import case. This commit
also fixes an error in the bandwidth calculation where the total number
of bytes was used instead of only the amount of bytes collected since
the last message was uploaded.