MXS-4657: Return object mismatches from the REST-API
Previously when an object of one type was given for an endpoint of
another type, the REST-API only reported a 404 Not Found error. This
made it hard to figure out why an endpoint that previously was valid
doesn't work with the given value. Usually this is either a typo in the
object name or a mixup of object types (e.g. `show service` instead of
`show server`). Another common case was with the module commands where
the order of the arguments is important but it's something that's easy
to forget.

By sorting the endpoints based on the broken down paths, the search
operation can identify the subset of endpoints that could match a given
request. This subset both speeds up all other comparisons as well as
makes it possible to report partial matches on requests. This results in
error messages of the form `<name> is not a <type>` which should make
errors easier to detect.

Now that the errors from the module search are logged in the response, a
new function that silences them is needed. This way only the errors
about the module (e.g. `X is not a module`) are logged.