Refresh the name of the tab used for altering the table
When the users choose the "Alter Table" option, a new query tab is
created with a name like "ALTER t1". The ddl editor in this tab
requires a connection, so it doesn't make sense to keep the name
when all other states of tab are reset when disconnecting the connection
without deleting that query tab.

This also moves the `filterEntity` helper function to the `Exteneder`
class.