MXS-4629 Handle deleting FKs
This simplifies the process of rendering rows. `rows` will be
a computed property which reacts on the changes of `stagingKeys`.

The `fetchReferencedTablesData` method now stores its value in
`newLookupTables` object keyed by referenced schema name and
table name. By doing this, there is no need to call `transformKeys`
to replace the referenced target names with ids. This helps to
ease to process of updating the two-way data bindings on keys.