Typecasting `TYPE` input in `ddl-editor`
The `TYPE` input uses `v-combobox` component which returns `null`
if the users delete all text in the input. This triggers an error
in the browser console about the `checkOptSupport` function.
That function expects the datatype to be a string but receiving `null`.

This converses `null` to empty string for `TYPE` input.