MXS-4629 Improve ERD UX
Apply the `changes-indicator` style, represented by an asterisk `*`,
to indicate modifications. If users make changes to a table but
close the DDL editor without executing the SQL, this indicator
will be added next to the altered table name.

Introduce the `new-tbl-indicator` for newly created tables.
This indicator is positioned next to the table name and will
persist until the SQL statement for table creation is successfully
executed.

This also adds `noDataText` props to `mxs-virtual-table-scroll-tbl`.