Migrate `sql-formatter`
Migrate to the latest version, which offers support for reserved
phrases such as `ON UPDATE SET NULL` and the ability to add a
customized dialect.

This installs `babel-plugin-transform-class-properties` package
to facilitate the migration and update the reserved keywords.

This also uses keywords and function names from the built-in `mariadb`
dialect.