Fix server status tracking
The lack of a status was indistinguishable from an empty status. By
expanding the stored type to 32 bits and reserving the maximum value for
a missing status, the problem is solved.

The EOF packets for COM_STMT_PREPARE are currently ignored by
MaxScale. The server status could be extracted from them if the
deprecate_eof protocol is not being used. Currently all state changes
that would've theoretically been caused by a COM_STMT_PREPARE are not
detected. Given the fact that preparing a SQL statement should have no
effect on the server status, ignoring it should be okay.

The TrxState::TRX_ENDING state was also sometimes set when a transaction
was about to start. This happened with `SET autocommit=1` statements
where the parser-based tracking correctly detected the state transition
but the server status based one considered it to be the end of a
transaction. This in turn caused the transaction to both start and end
at the same time that then caused a debug assertion to be hit.

This commit fixes a handful of tests that failed due to a missing server
status being misinterpreted as an empty server status.