Handle ignored replies in a separate function
This fixes a problem where the transaction replay state moves ahead too
early when a reply was not ignored when it should've been. This happens
when a backend that's unrelated to the transaction sends a response when
the interrupted query is being replayed.

Moving the handling of responses that should be ignored into a separate
function simplifies the checks needed to be done in the clientReply
function. This makes the rest of the code easier to understand as only
responses that change the observed state of the connection (i.e. ones
sent to the client) go there.

This fixes the sporadic failures of the mxs3924_sescmd_retry test that
started to fail when the implementation of MXS-4549 introduced the new
state where the partially executed query is processed.