Emulate old SUPER behavior in tests
The way that SUPER works now in MariaDB 11.0 is different from older
versions where it automatically gave the grants that were split from it
for 10.5 by MDEV-21743. Since we test with both new and old versions and
how the grants are given isn't important, the old behavior can be
emulated by granting all the grants to the monitor user.