MXS-4629 Export diagram as jpeg
The ERD diagram consists of `div` elements and `svg` element, there's
not built-in options for exporting it.

There are some libraries like `html2canvas, rasterizeHTML, dom-to-image`
but only `html2canvas` supports most modern browsers.

This installs `html2canvas` library to convert html to canvas and
then uses existing function to export it as jpeg.