MXS-4629 Fix `genScript` method
Script for altering foreign keys of created tables aren't generated
properly. In addition, there is an error about `Duplicate key on write or update`.

This improves `buildForeignKeySQL` method to create `ALTER TABLE`
statement for dropping the modified keys first before using the
existing key name for creating a new one.

This switches to only skip FK creation for new tables because
if there are any changes to the referenced tables, those FKs
should be added last. However, for FKs of created tables, it can be
included in the `# Alter tables` section.

This also moves script for altering created tables to the top.