MXS-3735 Adjust entity options
Remove the icons and keep only the text as the new labels are
more descriptive.

There will be 3 options:
Alter Table: For the new table, it's labeled as `Edit Table`.
Remove from diagram: Remove the table from the diagram.
Drop Table: Only for the created table.

This also renames some helper functions to more descriptive names.