MXS-3735 Fix redo behavior for blank ERD
In a blank ERD, if the users create a new table then undo and redo
the changes, the diagram `panAndZoom` values are broken because they
are all set to `NaN`. This happens because `getGraphExtent` returns
`undefined` when there are no nodes.

This also sets `test` as the default schema for blank ERD.