MXS-3735 Add a watcher for dispatching `queryDdlEditorSuppData` action
The action is needed to get supplement data for the `mxs-ddl-editor`.
e.g. engines, charsets and collations data.
When the users access the ERD worksheet with no connection bound to
it. i.e. after logging out and logging into the GUI, the above action
is dispatched in the `created` hook which is triggered once.

This adds a `watcher` for `activeErdConnId` to dispatch the
mentioned action. The action has an empty check before sending
the queries, so using the connection id as a watcher will be fine.