MXS-4682 Extend `mxs-graph-nodes` to emit `click` and `click-out-side` events
Clicking on a node will show reference points. The reference points
will be used for establishing FK.

The `click-out-side` event is added on the `document` scope as node
elements are positioned as `absolute`. Using existing directive from
vuetify won't work. Therefore, the event is emitted only if a node
was clicked.