Install Java with MDBCI
This should fix the Kafka installation work on Ubuntu Jammy.

Also fixed the test_kafkacdc test to not be so sensitive to timing of
how the replication is started. Previously the `RESET MASTER` could be
sent after MaxScale starts up which would possibly truncate the binlogs
in MaxScale and prevent all future events from being received.