Fix compiler warnings from GCC 13
The virtual to_json() defined in the Param base class was hidden by the
virtual to_json(value_type) function in the ParamNumber and ParamString
classes. A using declaration of the base class virtual function makes
them work the way they were intended to work.

In C++20, requires is keyword and GCC13 warns of this.