MXS-4682 Draw the referencing link in `ref-points`
This utilizes the `createShape` method from `utils` to draw a link
from the referencing point. This link will be removed after the
dragging event is ended.

This adds `INTERSECT_LEFT` to `TARGET_POS` as the reference
points are both placed on the left and right sides of the source node.

This also renames `svgGroup` variable to `linkContainer` as that
group element is used for containing only links.