MXS-4682 Add handlers for establishing an Fk
This adds `mouseenter` and `mouseleave` events to the entity column.
When the referencing link points to that specific entity column,
it automatically assigns the necessary data to refTargetData.

Upon emitting the `draw-end` event, the `onEndDrawFk` method handles
the creation of a new FK and emits the `on-create-new-fk` event.
The container component utilizes this event to add the FK to the
entity and initiates a diagram redraw.