MXS-3735 Fix FK alteration
Deleting a column which is being used as an FK column won't
work even if the users drop the FK first as the order of changes to
a table is not tracked. Therefore, the SQL for dropping FK must be
placed before any modifications to columns.

This splits `buildForeignKeySQL` function into smaller functions
that generate the SQL for dropping FKs and adding new FKs.