MXS-3735 Move the existing table's FK statements to the end of the script
When establishing a foreign key, the referenced column must be indexed.
The next commit will generate a statement for the referenced table
to create an index for the corresponding column. Order of changes to
tables is not tracked, therefore, if the users use the `Apply script`
button for executing script for all tables in the ERD, FK won't be
correctly formed.

This moves FK statements for the existing tables to the end of the
script to make sure the referenced column is properly altered before
establishing FK.

This also fixes `hasChanged` computed property to track also removed
nodes.