MXS-4682 Replace `findKeyTypesByColId` with `genColKeyTypeMap`
Creating a map for storing information about the key types of columns
is more efficient than finding key types for a column as the former
can be cached via computed properties and vuex getters.

The `colKeyTypeMap` getter in the `erdTasks` state module stores all
columns key types of all tables in the ERD which will be later used
for checking the referenced col key type.