MXS-3735 Fix `Remove from diagram` node option
The option has a conflict with adding a new FK feature.
When the users choose the mentioned option, the node is only removed
in the diagram but not from `stagingNodes`. This was an intention to
make sure the FK columns in other tables won't be affected. However,
when adding a new FK, the diagram is then updated with `stagingNodes`
which renders the removed node.

This adds `hidden` attribute to the node to prevent the node from
being rendered as well as links from being generated.