MXS-4613: Show correct info in SHOW SLAVE STATUS
The file and position is now correctly shown based on the information
being replicated instead of the local file information. The relay log
information now shows the old local file information.

Added a test case that makes sure that the binlogrouter returns the same
log name and position that MariaDB returns.