MXS-4682 Auto generate a plain index for referenced column
If the referenced column is not indexed, it should automatically
generate a plain index for that column. This is sensible when
creating a new FK via `ref-points` component as it's in the context
of ERD. On the other hand, if the users create a new FK via the
ddl editor, it shouldn't generate one as it involves altering the
referenced table.

This improves `TableScriptBuilder` to build SQL for all keys.