MXS-4682 Validate FK key column and referenced column
"The foreign key columns and the referenced columns must be of the
same type, or similar types. For integer types, the size and sign
must also be the same."
https://mariadb.com/kb/en/foreign-keys/#syntax

This commit does the following:

* Validates same type and integer type cases.

* Removes `SQL/PL` data types as the ddl editor uses
the result of `SHOW CREATE TABLE` to parse table which automatically
returns mariadb data types regardless `SQL_MODE`.

* Removes the optional size wrapper. i.e. `()`.