MXS-3735 Persisting staging nodes
As a user, making an ERD and then accidentally closing the browser
without executing the script, all the new tables or changes are lost
which is a bad UX.

This persists the staging nodes in `ErdTask` model to keep
data in IndexedDb.

If the users generate ERD from the existing tables, the initial parsed
tables are stored in `tables` field.