MXS-4682 Add option to quickly change entity relationship
This adjusts `mxs-sub-menu` component to act as a wrapper component
of `v-menu` to make it more customizable.

A FK link can have either `Change to One To One` or `Change to One to Many`
options. If it's already 1 to 1, the other option will be shown in
the context menu.

This also adds `click` event to FK link in the diagram. Clicking
on the link will also open the context menu.