MXS-3735 Auto detect identifying and non-identifying relationship
"An identifying relationship will have a solid line (where the PK
contains the FK). A non-identifying relationship is indicated by
a broken line and does not contain the FK in the PK"

This adds `isIdentifyingRelation` function to do so. If it returns
true, the styles property of the link ojbect will have `dashArr` style
set to 0.