MXS-3735 Fix `buildAlterScript` method
Toggling a UQ column returns invalid extra comma as the
`hasColDefChanged` is true when the UQ value changes. However,
Unique key values was modified to create SQL in the
`buildOtherKeys` method, not `buildColDefAlterSQL`.

This also fixes context menu in `diagram-ctr`. The menu should
be shown only when there is `activeCtxItemId`