MXS-3735 Parse keys of a category into a map
Dealing with an array of keys within a category proves to be more
troublesome than advantageous. Unlike the order of the column in
a composite key, the order of keys in a category is not matter.
It causes big trouble to get the diff between initial keys and
staging keys when inserting and removing the keys.  By converting
it into an object map, the diffs can be generated accurately without
worrying about the order of keys in a category as the order of
js object properties are not matter.