MXS-3991 Facilitating the implementation of `index-cols-list` component
Some computed properties and methods can be made reusable for
the upcoming component .i.e `index-col-list`.

This removes the watcher for `row` changes in `mxs-virtual-scroll-tbl`
as the reset of the `selectedTblRows` usually should be done by the
parent component. In general, the whole component should be re-rendered
which is the case for the `entity-editor-ctr`.
This also extends the component to emit row click event.