MXS-4690: Truncate partially written transactions
By truncating the latest binlog file to the last known good transaction
on startup, the need to buffer transactions in memory is removed. This
makes it possible to replicate transactions that would not fit into
memory. It should also make the replication easier on the CPU as it
reduces the amount of memory copies that are made.

An alternative implementation would've been to write the transactions
into a small temporary file that would then get concatenated to the
binlog file but the APIs required to efficiently append to files
(copy_file_range) are not available in all kernel versions and this
implementation would still end up writing the events twice. The
optimistic approach with truncation should end up being the simpler and
thus the more efficient solution.

The ddl_only mode in 22.08 will need additional work as it currently
relies on the transaction buffer to discard transactions that aren't
DDLs. By storing a flag for whether a write to the file should be done
or not should solve it and also make the implementation of it faster.