MXS-4696: Fail fast when permanently broken
If the backends failed in a permanent manner, the retrying should not be
done. This causes the failures to happen much faster in situations where
something is very broken. Failing faster in these cases reports the
errors back to the client faster which should speed up figuring out why
it fails in the first place.