MXS-3735 Parse columns into a map
If the users edit indexes via the `index-definitions` editor component,
the changes are updated into `keyCategoryMap`. As a result, the
transformed data used in the `col-definitions` component is out of
sync because it is done once right after parsing the table.

Instead of persisting 2d array of columns, it's better to persist
the parsed data as an object. It was converted to 2d array because
`mxs-virtual-scroll-tbl` reads that data structure. However, the
transformation can be done in `col-definitions` component via
a computed property. By doing this, any changes to `keyCategoryMap`
will trigger a rerun on the `transformedCols` computed property.

This also renames several variables and properties to more meaningful
names.