MXS-3735 Fix <keep-alive> issue
Watchers in `table-header` component of the `mxs-virtual-scroll-tbl`
component are triggered when switching between worksheets. As a result,
it fails to compute the width of the header. The watchers are
triggered because the workspace keeps alive worksheet components.
Data is chosen based on active id, therefore, the watchers must be
removed when components are deactivated.

This adds watchers of `table-header` on `activated` and remove
on `deactivated` hooks.

This also removes obsolete components.