MXS-4637 Replace param region with region_name and region_oid
In the region aware Xpand 23.08 a region has both an oid and
a name, and both need to be unique. Both of them must be unique,
but the oid presumably cannot be changed while the name presumably
can. The oid is system created long integer, i.e. human unfriendly,
while the name is an admin specified string, i.e. human friendly.
Better leave to the admin, which one he wants to use.