MXS-3991 Caching alter-table-editor staging data in `QueryTabTmp` model
There is a glitch in an edge case where there are more than two tabs
for altering table. If the users delete the inactive tab, the `stagingData`
of the `alter-table-editor` component of the active tab is overwritten by
the inactive tab.

This moves the stagingData to `QueryTabTmp` model, the field name is
`staging_alter_data`.

This also makes `file-dlg-ctr` component reusable at worksheet scope
and prompt the dialog if the users delete tab using for altering
table.