MXS-4685: Avoid tellg() usage
Using gcount() to extract the number of characters read is significantly
faster than using tellg() to get the current position.

For encrypted events, the actual size on disk might be slightly larger
than it is when stored in memory. This necessitates the storage of the
on-disk size as it is needed to calculate the next read position.

Testing with a NVMe drive with about 400MiB of binlogs, the reading is
roughly seven times faster with gcount() than it was with tellg(). The
odd thing is that the speed is the same when a tmpfs filesystem is used
and the data is completely in memory. This means that the performance of
the event reading was CPU bound like was seen in practice where the
bottleneck was identified.