MXS-3735 Add icon button to toggle relationship highlight
It's nice to have this feature especially when the
`Draw Foreign Key Links to Associated Columns` mode is not in used.
The users can toggle the highlight for all links in the ERD.

This improves `GraphConfig` class to have a simpler way to
update its config.

This also fixes a minor style issue with `erd-key-icon` by
assigning fixed width to it.