MXS-3735 Generate ERD nodes into a map
Object property access has an average time complexity of O(1) for
retrieval, while retrieving item in an array has a time complexity
of O(n). The order of ERD nodes is not a matter, so it's completely
fine to convert it into an object map. Retrieving and deleting
the node would be a lot faster and simpler.