MXS-3991 Automatically lazy-loading and parse new referenced table
The `alter-table-editor` has 1 table in the `lookupTables` props,
so when creating or modifying an FK, it can only reference to itself
or to tables that have been referenced.

This enhances the `mxs-ddl-editor` component to accept hinted
reference targets which is a list of tables that are in the same
schema as the table being altered/created. It could be from
other schemas as well but for now the `alter-table-editor`
will pass down only tables in the same schema.