MXS-4701: Backport Server GTID changes to 2.5
The change was originally done for 23.08 where the GTID positions were
sent to the RoutingWorker threads directly instead of being assigned via
the MainWorker with WorkerGlobal::assign(). The key benefit of this is
that it no longer involves the MainWorker which frees it up to serve
REST-API clients.

This commit includes an improvement to the original commit where the
value on the MainWorker is also updated asynchronously. This retains the
behavior of the WorkerGlobal where the values are up-to-date on the
MainWorker but without the synchronous nature of WorkerGlobal::assign().