Fix and test ddl_only mode
Added a test case for the `ddl_only` mode and fixed the feature to work
with the new way in which transactions are handled in the binlogrouter.

The MASTER_GTID_WAIT error now also prints out which node the command
failed on. Since it only makes sense to run it on either MaxScale or the
slave, the case where it's executed on the master can be ignored.

Also fixed two formatting errors in writer.cc and added information into
the debug log messages for detecting when events are not written to the
binlog file.