Store PS info in std::vector
The number of prepared statements is likely to be small and the IDs are
likely to be contiguous. This means that storing them in a std::vector
gives better locality. This seems to reduce the time spent in the code
that previously used an std::unordered_map by roughly half.