MXS-4685: Seek GTIDs incrementally
The existing GTID exclusion mechanism can be used to implement
incremental seeking into the binlogs by always starting from the
beginning of each file. The only thing that needs to be done is to push
down the timer into the fetch_event() call so that it knows to stop if
it was taken too long. This solves the problem where `maxctrl list
servers` was unresponsive while a `START SLAVE` command was being
executed.

The correct file to start from still needs to be identified which means
a small amount of blocking work still remains. As the amount of data
read is comparatively small, less than a thousand bytes per binlog file,
the amount of blocking appears to be acceptably small.

This approach of incremental seeking is noticeably slower than doing the
search in the search_file function. This is likely something that can be
improved but the immediate problem of the worker's message queue being
blocked is now solved. Compared to the original performance, this will
still end up being faster mainly due to the lack of tellg() calls that
it used to do.