Fix mxs2057_systemd_watchdog
Now that the luafilter is no longer available when testing and with the
reduction in regex matching limits, the old test configuration is not
slow enough for 23.08. This can be remedied by tuning the SQL to fall
just below the matching limit and by using up all the patterns as well
as creating multiple copies of the filter. This slows it down enough
that the watchdog is now again reliably triggered.