MXS-4506 Deliver backend error packet to client directly
If backend authentication results in an error, the session could
be closed already in the backend event handler. This means that
a separate client-side event may never be ran even if manually
triggered.

Handle this by sending the backend reply to client in the backend
event handler. A client dcb event is still triggered to advance
the state machine in case session survives.