MXS-4506 Return data from SSL read even if an error occurs
SSL read operation can successfully read data even if a later
read in the loop fails due to fatal error. This commonly happens
if server rejects authentication: in this case the server sends
the error packet and immediately closes the connection. To process
the error packet, socket_read_SSL() must first return success and
return the io error as a separate event.

Also, added a new function to log OpenSSL errors, as the previous
function is unclear in usage.