Use a private TrxTracker in QueryClassifier
This removes the need for the second update_route_info() function which
used the session's protocol information for the transaction state
information. In the client protocol this was not working as intended
since the information was used before the transaction state was updated
to represent the query actually being classified. Also, the transaction
state in the client protocol was not updated at all if the session did
not have the transaction tracking capability. This meant that the
QueryClassifier in readwritesplit and the one in mariadb_client.cc
behaved slightly differently.

Moving the TrxTracker inside the QueryClassifier also makes the usage of
the QueryClassifier simpler by bundling all the functionality in one
class.

If both transaction tracking and session command history are needed
(e.g. cache filter with readwritesplit), the transaction state is copied
over to the one in the protocol data. This avoids having to do the
tracking twice.

This commit also removes the need for the transaction tracking
capability in the smartrouter as the same information is now available
in the QueryClassifier in a more accurate form.