MXS-3986 Make binlog.index impervious to change
Self healing index. If deleted it is immediately recreated,
if modifed it is restored on the next directory change
(a file is created or deleted in the directory).

The binlog.index contents is not used by maxscale but it is a
requirement to maintain it.

Funny detail. Most editors will create a temporary backup file while
a file is being edited. This causes pinloki to restore the index
as soon as the temporary file is deleted.