MXS-4674: Don't use semi-sync in replicator
It should be disabled as anything that uses the replicator does not
actually store the information and only processes it. If this were to be
done, there should be a parameter that enables it instead of it being
done automatically.

It's likely that the sporadic failures of mxs1687_avro_ha are caused by
the problem described in CONC-659. If the is_semi_sync flag is set once
per 256 seconds, the test uses the wrong offset into the event which in
turn causes the wrong event type to be read. This causes a read past the
end of the vector that contains the metadata information.

With the removal of the semi-sync, the tests still fail but this time in
a similar manner to what is described in MXS-4668: the server will close
the connection when the connector sends an unexpected semi-sync ACK
packet.