MXS-3991 Add `Show Create` option
For schema and table nodes, DDL can be viewed by selecting the `View Insights`
option. However, for view, trigger, procedure and function nodes, the
`SHOW CREATE` command returns other info along with the DDL, so it needs
a special handler.

This adds `CREATION_INFO` spec which will be shown only for
`view, trigger, ...`