MXS-4506 Improve responsiveness on backend authentication error
This improves responsiveness with errors that occur on the
MaxScale side and have no associated error packet. These are
typically caused by misconfiguration and/or differences in user
account information between MaxScale and server. Calling the
error handler causes the router to consider the backend connection
broken instead of waiting for the server to disconnect.