MXS-4668: Use Connector/C 3.3.6 prerelease tag
The tag contains the option that allows semi-sync replication to be
conditionally enabled. This also removes the need to manually set the
user variable which is an improvement in code readability.

The `rpl_semi_sync_slave_enabled` parameter for the binlogrouter now
directly controls this new connector option.

The other modules that replicate from MariaDB do not yet support
semi-sync as they aren't capable of storing information and thus do not
make the data durable.