Backport mxs::Reply::describe()
The functionality is extremely useful for tracking what the response
from the server was. Having the actual reply type in the info logs has
in practice proven to be helpful enough that backporting it into 2.5 is
smart.

This commit also adds log messages for readconnroute
responses. Previously only queries were logged but this left out
information that revealed whether there were multiple pending queries or
not.