Fix duplicate creation of BinglogIndexUpdater
A call "new BinglogIndexUpdater()" previously moved mysteriously
reappeared in its original spot in the code with a commit of mine.
Remove that and rename BinglogIndexUpdater to BinlogIndexUpdater.

The bug is not critical as the extra updater does not cause any
havoc. It may not even start as it was moved in an even earlier
commit to prevent it from starting before the binlog directory
is created.