Fix lower_case_table_names documentation
The numbers appear to get converted into a numbered list which causes
the values to be wrong. Using a bulleted list with the values as code
blocks prevents this.

Also added the missing information that the behavior of the parameter
changed in a recent release. This is important as in older releases the
behavior of `lower_case_table_names=1` was not as it is now documented.