MXS-4717 Invalidate information_schema as needed
SELECTs targeting information_schema will be cached as any
other SELECT. However, that table will not change due to
INSERTs but as a side-effect of e.g. CREATE TABLE and
DROP TABLE.

A straightforward way to deal with this is to add, when a table
is CREATED, DROPPED, etc., relevant information_schema tables
to the set of tables whose cache entries should be evicted.

This is something of an edge-case, but with this change certain
NoSQL tests pass also when caching is enabled.