Reduce memory consumption usage in `col-definitions` component
If a table has a large number of columns, several vuetify inputs
will be rendered at once in the table. Though virtual scroll
is used to improve performance but there is still a delay when
switching between altered table query tabs. This is because
vuetify input components internally have several computed
and reactivity data which are all kept in memory. Particularly,
when rendering the datatypes, charset and collation inputs, the
same items are being kept in memory unnecessarily.

This adds `mxs-lazy-input` component which renders a readonly
input until the users click or focus on the component, the
actual vuetify input component will be rendered. When the users
finish the changes with the input, it switches back to the readonly
input which takes less memory.