23.08.0 Replace LICENSE2302.TXT with LICENSE2308.TXT
This follows the earlier approach, but this should suffer from
the same problems as the version file. So, there probably ought
to be separate license files for every major release. For later.

The change date is now the same as that of the previuous 23.02
release. Simpler that way and a difference of one month is
irrelevant in the big scheme of things.