Document common monitor parameters uniformly
The monitor_interval example used a slightly different value than what
was set as the default value which made it very easy to confuse it for
the default. The actual default value was also documented twice.

Converted the documentation to the format with a fixed "header" that
specifices the type, the default and whether it can be modified at
runtime.