MXS-4716: Add session_trace_match
Currently the in-memory log messages that session_trace collects are
only written to disk in situations that are caused by internal MaxScale
errors (bugs) or if the client connection is abruptly closed. The new
parameter can be used to control if the messages are written to disk
when the session ends.

The way the session_trace log is presented is backwards from the log
file's point of view: newest log messages are at the top and the oldest
ones are at the bottom. This is the inverse of how log messages are
usually written in MaxScale and should be modified to write the messages
in the correct order.