MXS-4716: Reorder session_trace output
The output is now logged in chronological order with the oldest messages
first and newer ones last. The messages are also logged as separate log
messages to prevent the truncation that would otherwise happen. It also
keeps the format of the log consistent while also keeping the log
timestamps in increasing order. Previously the trace log dump could
introduce messages that had older timestamps.