MXS-3735 Simplify charset and collation data input flows
This removes the need for storing the column charset if
its value equals to table charset. e.g. When the collation
is not the default value, SHOW CREATE command includes the
charset value in the result.
```
CREATE TABLE
  `t1` (
    `meta` LONGTEXT CHARACTER SET utf8mb4 COLLATE utf8mb4_bin DEFAULT NULL,
  ) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4
```
In the above case, if the `skipColCharsetIfEqual` parameter is true,
`charset` of the column will be set to `undefined`.

Column with default charset and collation will have `DEFAULT`
as an input placeholder.

This also adds an equal check for some lazy inputs to prevent
unnecessary updates. In addition, this replaces `stringifyClone` in
`mxs-virtual-scroll-tbl` component with `lodash.cloneDeep`
to preserve `undefined` value.