Only check the latest completed command
In History::Subscriber::add_response(), there's no need to check the
other pending responses as they'll be processed in the callback function
that is called once a new response that is accepted arrives. For the
case where the history is already executed and a reconnection is taking
place, this completely avoids allocating the memory that was previously
used to temporarily store the ID of the command that was just executed.

The ID and the result of the command that triggered the call is now
given to History::Subscriber::compare_responses(). This avoids having to
iterate over the elements to look for completed responses where a simple
lookup would suffice.

Renamed the MXS_SESSION::is_started() function to
MXS_SESSION::is_alive() since it better describes the intended use of
the function. Since the number of uses of this function doubled with
this commit, it seemed like an appropriate time to rename it.

This commit also fixes a minor problem in the info logging done in
MariaDBBackendConnection::read_history_response() where the logged
command ID was always zero. This happens because add_response() always
resets it.