Fix `busy connection` error while querying ddl editor supplement data
When the users alter a table the first time, `queryDdlEditorSuppData`
action is dispatched twice asynchronously. `onAlterTable` method is
supposed to dispatch the `handleAddQueryTab` action which creates a
connection and binds it to the `ALTER_EDITOR` query tab. After that,
it dispatches `queryDdlEditorSuppData` and queries table DDL. However,
after adding the `ALTER_EDITOR` query tab, there is another call to
dispatch `queryDdlEditorSuppData` in the `queryEditors/handleInitialFetch`
action which is triggered by the `activeQueryTabConn` watcher in
the `QueryEditor` worksheet.

This adds a data check for `handleInitialFetch` before dispatching the
action to prevent duplicated dispatches.