Set wait_timeout in mxs3663_universal_causal_reads
The test uses a transaction to insert a rows into the table which has a
primary key. If the connection to the server is blocked after the row is
inserted, the monitor in MaxScale will cause the connection to it to be
terminated. Since the blocking is done by preventing network
communication between MaxScale and the database, it's possible that the
database does not notice that the connection which holds the locks is
now dead. When the test is run with transaction_replay, the stale
transaction will end up blocking the replayed transaction.

To prevent this from happening, wait_timeout is set to half of what the
main thread waits before blocking a new node. This should cause the
partial transaction to be rolled back and the next time the transaction
is replayed, the row can again be inserted.