MXS-4732: Process signals outside the signal hander
The use of Worker::execute() is not safe in a signal handler as it
causes a CustomTask class to be allocated from the heap which is not
signal-safe. This could be fixed by using the low-level post_message()
function but it would not work for the EventMessageQueue that's used in
newer versions.

To solve both cases, a pipe in the MainWorker can be used to write a
function that will then get executed once epoll notices that there's
data in the pipe. Given the very narrow scope of this need, this can be
implemented completely in the MainWorker without having to make it a
generic mechanism in the Worker base class.