MXS-4725: Add debug stacktrace endpoint
The stacktraces are largely similar to what mxb::emergency_stacktrace
produces except that the values are broken down into JSON arrays for
easier processing.

The stacktraces are collected using real-time signals which behave
slightly differently than normal signals where they are guaranteed to be
delivered as many times as they are sent. This also avoids using a
predefined signal for what amounts to internal use which is what the
real-time signals are designed for. All currently supported versions of
Linux have real-time signals.

The profiling is done in a way that affects the actual worker threads in
a minimal manner. Further post-processing of the stacktraces is done
asynchrononously by the MaxScale thread pool which means the MainWorker
is only blocked for a short while when the data is being collected. The
act of demangling the symbols is somewhat costly and offloading this to
the client is the most sensible thing to do.